chore(warnings): progress commit - clear C4018, C4189, C4245, C4267, C4456 categories
Replaces -1 sentinels assigned to unsigned types with explicit static_cast<DestType>(-1) (177 sites total across model + controller + view + tests), promotes a few remaining int loop counters to uint32_t, removes verified-dead locals, and renames inner shadowing locals where the inner/outer types differ. Also rewrites INVALID_BRANCH_LEVEL to drop a stray trailing ; and uses EventManager::InvalidSubscriptionToken instead of a magic -1 in TimelineArrow. No public API changes; build clean, all five categories at zero. Updates WARNINGS.md to reflect live state.

+++ src/controller/src/system/rocprofvis_controller_table_system.cpp
@@ -495,7 +495,7 @@ SystemTable::UnpackArguments(Arguments& args, QueryArguments& out) const
     std::vector<const char*> string_table_filters_ptr;
     bool summary = false;
     rocprofvis_controller_track_type_t track_type = kRPVControllerTrackTypeSamples;
-    uint64_t table_type = kRPVControllerTableTypeEvents;
+    uint64_t table_type = static_cast<uint64_t>(kRPVControllerTableTypeEvents);
 
     result = args.GetUInt64(kRPVControllerTableArgsType, 0, &table_type);
     if (result == kRocProfVisResultSuccess)
